requirements.txt
attrs==20.3.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
distro==1.6.0
Django==2.2.24
entrypoints==0.3
gunicorn==20.1.0
httplib2==0.20.2
hyperlink==21.0.0
importlib-metadata==4.6.4
incremental==17.5.0
jeepney==0.7.1
Jinja2==3.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.3.0
launchpadlib==1.10.15.1
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
macaroonbakery==1.3.1
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.1.1
pexpect==4.8.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycrypto==2.6.1
PyHamcrest==2.0.2
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5b0
python-dateutil==2.8.1
python-debian===0.1.39
pytz==2021.3
PyYAML==5.4.1
requests==2.25.1
requests-unixsocket==0.2.0
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.11
Twisted==20.3.0
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0

django-heroku
dj-database-url
psycopg2-binary
whitenoise

pandas==1.3.5
numpy==1.22.0
sklearn==0.0
spotipy==2.19.0
tensorflow-cpu==2.7.0
